requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
streamlit-extras
altair<5
googletrans==3.1.0a0
google.generativeai
speech_recognition
soundfile
sounddevice
langchain